requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
supabase>=2.18.0
pydantic>=2.6.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
httpx>=0.25.0